seeder complete

+++ app/Http/Controllers/Backsite/ConsultationController.php
@@ -14,7 +14,7 @@ use App\Http\Requests\Consultation\StoreConsultationRequest;
 use App\Http\Requests\Consultation\UpdateConsultationRequest;
 
 // use everything here
-// use Gate;
+use Illuminate\Support\Facades\Gate;
 use Auth;
 
 // use model here
@@ -39,6 +39,7 @@ class ConsultationController extends Controller
      */
     public function index()
     {
+        abort_if(Gate::denies('consultation_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $consultation = Consultation::orderBy('created_at', 'desc')->get();
 
         return view('pages.backsite.master-data.consultation.index', compact('consultation'));
@@ -78,6 +79,7 @@ class ConsultationController extends Controller
      */
     public function show($id)
     {
+        abort_if(Gate::denies('consultation_show'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         return view('pages.backsite.master-data.consultation.show', compact('consultation'));
     }
 
@@ -89,6 +91,7 @@ class ConsultationController extends Controller
      */
     public function edit(Consultation $consultation)
     {
+        abort_if(Gate::denies('consultation_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         return view('pages.backsite.master-data.consultation.show', compact('consultation'));
     }
 
@@ -116,6 +119,7 @@ class ConsultationController extends Controller
      */
     public function destroy(Consultation $consultation)
     {
+        abort_if(Gate::denies('consultation_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $consultation->forceDelete();
         alert()->success('Success Message', 'Successfully Delete Consultation');
         return back();
